requirements.txt
numpy==1.16.3
pandas==0.24.2
opencv-python==4.1.0.25
scipy==1.2.1
tensorflow-gpu==1.13.1
keras==2.2.4
scikit-learn==0.21.1
scikit-image==0.16.2
image-classifiers==1.0.0b1
matplotlib==3.0.3
pyntcloud==0.1.2
pythreejs==2.1.1
IPython
